# services/api/Models/Entities.cs
namespace Shancrys.Api.Models;

public class User
{
    public Guid Id { get; set; }
    public Guid TenantId { get; set; }
    public required string Email { get; set; }
    public required string PasswordHash { get; set; }
    public required string Name { get; set; }
    public List<string> Roles { get; set; } = new();
    public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
    public DateTime? LastLoginAt { get; set; }
}

public class Project
{
    public Guid Id { get; set; }
    public Guid TenantId { get; set; }
    public required string Name { get; set; }
    public required string Location { get; set; }
    public string? Description { get; set; }
    public ProjectStatus Status { get; set; } = ProjectStatus.Planning;
    public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
    public DateTime UpdatedAt { get; set; } = DateTime.UtcNow;
}

public enum ProjectStatus
{
    Planning,
    Active,
    Paused,
    Completed
}

public class ModelVersion
{
    public Guid Id { get; set; }
    public Guid ProjectId { get; set; }
    public required string FileName { get; set; }
    public required string FilePath { get; set; }
    public long FileSize { get; set; }
    public required string FileHash { get; set; }
    public required string Format { get; set; }
    public FileType FileType { get; set; }
    public string? Description { get; set; }
    public int Version { get; set; }
    public DateTime UploadedAt { get; set; } = DateTime.UtcNow;
    public ModelStatus Status { get; set; } = ModelStatus.Processing;
    public Dictionary<string, object> Stats { get; set; } = new();
    public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
    public Guid CreatedBy { get; set; }
    public Project Project { get; set; } = null!;
}

public enum FileType
{
    IFC,
    DGN,
    RVT
}

public enum ModelStatus
{
    Uploaded,
    Processing,
    Ready,
    Failed
}

public class Element
{
    public Guid Id { get; set; }
    public Guid ModelVersionId { get; set; }
    public Guid VersionId { get; set; }
    public required string IfcGuid { get; set; }
    public required string Guid { get; set; }
    public required string Name { get; set; }
    public required string Type { get; set; }
    public required string Discipline { get; set; }
    public string? Level { get; set; }
    public double[]? BoundingBox { get; set; }
    public double? VolumeEstimated { get; set; }
    public decimal? CostEstimated { get; set; }
    public Dictionary<string, object> Properties { get; set; } = new();
    public Dictionary<string, object>? Attributes { get; set; }
    public ModelVersion ModelVersion { get; set; } = null!;
}

public class Activity
{
    public Guid Id { get; set; }
    public Guid ProjectId { get; set; }
    public required string Wbs { get; set; }
    public required string WbsCode { get; set; }
    public required string Name { get; set; }
    public string? Description { get; set; }
    public DateTime PlannedStartDate { get; set; }
    public DateTime PlannedEndDate { get; set; }
    public DateTime StartPlanned { get; set; }
    public DateTime EndPlanned { get; set; }
    public DateTime? ActualStartDate { get; set; }
    public DateTime? ActualEndDate { get; set; }
    public DateTime? StartActual { get; set; }
    public DateTime? EndActual { get; set; }
    public int Duration { get; set; }
    public List<string> Predecessors { get; set; } = new();
    public decimal ProgressPercent { get; set; } = 0;
    public double Progress { get; set; } = 0;
    public ActivityStatus Status { get; set; } = ActivityStatus.NotStarted;
    public Dictionary<string, object> Metadata { get; set; } = new();
    public Project Project { get; set; } = null!;
}

public enum ActivityStatus
{
    NotStarted,
    InProgress,
    Completed,
    OnHold
}

public class ElementActivityMapping
{
    public Guid Id { get; set; }
    public Guid ElementId { get; set; }
    public Guid ActivityId { get; set; }
    public MappingType MappingType { get; set; } = MappingType.Manual;
    public decimal Confidence { get; set; } = 1.0m;
    public double Weight { get; set; } = 1.0;
    public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
    public Element Element { get; set; } = null!;
    public Activity Activity { get; set; } = null!;
}

public enum MappingType
{
    Manual,
    AutoGenerated,
    AIAssisted
}

public class ProgressRecord
{
    public Guid Id { get; set; }
    public Guid ActivityId { get; set; }
    public Guid? ElementId { get; set; }
    public DateTime Timestamp { get; set; } = DateTime.UtcNow;
    public ProgressType Type { get; set; }
    public double? Value { get; set; }
    public string? MediaUrl { get; set; }
    public Dictionary<string, object>? GeoLocation { get; set; }
    public Guid Author { get; set; }
}

public enum ProgressType
{
    Percentage,
    Photo,
    Inspection
}

public class CostRecord
{
    public Guid Id { get; set; }
    public Guid ProjectId { get; set; }
    public Guid? ActivityId { get; set; }
    public DateTime Date { get; set; }
    public required string Category { get; set; }
    public decimal Value { get; set; }
}