# specs/api-openapi.yaml
openapi: 3.1.0
info:
  title: Shancrys 4D Platform API
  version: 1.0.0
  description: |
    API REST para plataforma de construção digital 4D.
    Integra modelos BIM com cronogramas para simulação temporal de obras.
servers:
  - url: http://localhost:5000/api/v1
    description: Desenvolvimento local
  - url: https://api.shancrys.com/api/v1
    description: Produção
security:
  - BearerAuth: []
...
paths:
...
  /projects:
    get:
      tags: [Projects]
      summary: Listar projetos do tenant
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [planning, active, paused, completed]
      responses:
        '200':
          description: Lista de projetos
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
                  total:
                    type: integer
                  page:
                    type: integer
                  pageSize:
                    type: integer
    post:
      tags: [Projects]
      summary: Criar novo projeto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: Projeto criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 500
...
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        name:
          type: string
        location:
          type: string
        status:
          type: string
          enum: [planning, active, paused, completed]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProjectCreate:
      type: object
      required: [name, location]
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 200
        location:
          type: string
        description:
          type: string
...